fix: webconfig duplication
