+++ backend/src/main/java/com/SeeAndYouGo/SeeAndYouGo/config/WebConfig.java
@@ -1,14 +1,23 @@
 package com.SeeAndYouGo.SeeAndYouGo.config;
 
+import com.SeeAndYouGo.SeeAndYouGo.oAuth.jwt.TokenProvider;
+import com.SeeAndYouGo.SeeAndYouGo.visitor.VisitorInterceptor;
+import lombok.RequiredArgsConstructor;
 import org.apache.http.HttpHeaders;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
+@RequiredArgsConstructor
 public class WebConfig implements WebMvcConfigurer {
 
     public static final String ALLOWED_METHOD_NAMES = "GET,POST,PUT,DELETE,OPTIONS";
+    private final RedisTemplate<String, String> redisTemplate;
+    private final TokenProvider tokenProvider;
+
 
     @Override
     public void addCorsMappings(final CorsRegistry registry) {
@@ -17,5 +26,11 @@ public class WebConfig implements WebMvcConfigurer {
                 .allowedOrigins("*")
                 .allowedOriginPatterns("*")
                 .exposedHeaders(HttpHeaders.LOCATION);
+    }
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(new VisitorInterceptor(redisTemplate, tokenProvider))
+                .addPathPatterns("/**");
     }
 }
